Fix : 아기댓글 삭제 시 엄마댓글도 삭제되는 이슈 해결
Former-commit-id: 08c3a2b2b65706bd0b87f2b7d201e12c08e3a940

+++ BE/spring/paragraph-service/src/main/java/com/booklet/paragraphservice/service/CommentServiceImpl.java
@@ -93,7 +93,9 @@ public class CommentServiceImpl implements CommentService {
                 // 본인만 남았다면
                 Comment parent = commentRepository.findById(parentId).orElseGet(Comment::new);
                 if (parent.getCommentId() == null) return false;
-                commentRepository.delete(parent);
+                if(parent.getCommentContent().equals("")){ // 삭제된 댓글이면
+                    commentRepository.delete(parent);
+                }
             }
         }
